Add files via upload
This is an android mobile application.  The app lets the user decide his daily meals everyday for breakfast, lunch, and dinner.  User can add food items to list of foods and the user can use one of those food items later on as one of the things to eat any day for breakfast, lunch, or dinner.